==> scripts/02-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data from [...UPDATE THIS...]
# Author: Sandy Yu
# Date: 21 October 2024
# Contact: jingying.yu@mail.utoronto.ca
# License: MIT
# Pre-requisites: [...UPDATE THIS...]
# Any other information needed? [...UPDATE THIS...]


#### Workspace setup ####
library(opendatatoronto)
library(tidyverse)
# [...UPDATE THIS...]

#### Download data ####
# [...ADD CODE HERE TO DOWNLOAD...]



#### Save data ####
# [...UPDATE THIS...]
# change the_raw_data to whatever name you assigned when you downloaded it.
write_csv(the_raw_data, "inputs/data/raw_data.csv")
